+++ Project_SoCMainFrame/os_global.cpp
@@ -7,7 +7,7 @@
 Event_List event_list;
 
 //===========OS_STATUS==========
-int systime; //System Time in milliseconds
+long int systime; //System Time in milliseconds
 
 Task_List* thread_list[MAX_THREADS]={NULL}; //instantiate a task list	
 byte thread_count=0;	// Number of current task lists	
@@ -17,7 +17,7 @@ byte lib_init_count=0;	//how many inits?
 
 struct Task errptr={0,0,{0,0},0,0};
 
-int over_time;
+long int over_time;
 
 
 //===========Attached(Combine,Devide)=====
